refactor(projects): Move chart drawing to charts
